backend/requirements.txt
Django>=4.2.16,<5.0
djangorestframework>=3.15.2,<4.0
django-cors-headers>=4.6.0,<5.0
